🚚 Rename: domains/review 폴더 생성, `ReviewBadge` 이동
