+++ src/domains/review/types/badge.ts
@@ -1,4 +1,4 @@
-import BADGE from '@/shared/constants/badge';
+import BADGE from '@/domains/review/constants/badge';
 
 export type BadgeKindType = 'preference' | 'taste' | 'texture';
 export type BadgeShapeType = keyof typeof BADGE;
